+++ libevmasm/AssemblyItem.cpp
@@ -42,8 +42,8 @@ namespace
 
 string toStringInHex(u256 _value)
 {
-	std::stringstream hexStr;
-	hexStr << std::uppercase << hex << _value;
+	stringstream hexStr;
+	hexStr << uppercase << hex << _value;
 	return hexStr.str();
 }
 
@@ -158,7 +158,7 @@ size_t AssemblyItem::bytesRequired(size_t _addressLength, Precision _precision) 
 			return 2;
 	}
 	case VerbatimBytecode:
-		return std::get<2>(*m_verbatimBytecode).size();
+		return get<2>(*m_verbatimBytecode).size();
 	default:
 		break;
 	}
@@ -245,6 +245,18 @@ string AssemblyItem::getJumpTypeAsString() const
 	default:
 		return "";
 	}
+}
+
+void AssemblyItem::setJumpType(string const& _jumpType)
+{
+	if (_jumpType == "[in]")
+		m_jumpType = JumpType::IntoFunction;
+	else if (_jumpType == "[out]")
+		m_jumpType = JumpType::OutOfFunction;
+	else if (_jumpType.empty())
+		m_jumpType = JumpType::Ordinary;
+	else
+		solThrow(AssemblyImportException, "Invalid jump type.");
 }
 
 string AssemblyItem::toAssemblyText(Assembly const& _assembly) const
@@ -411,7 +423,7 @@ size_t AssemblyItem::opcodeCount() const noexcept
 	}
 }
 
-std::string AssemblyItem::computeSourceMapping(
+string AssemblyItem::computeSourceMapping(
 	AssemblyItems const& _items,
 	map<string, unsigned> const& _sourceIndicesMap
 )